@@ -1,4 +1,5 @@
-import { Observable } from 'rxjs/Observable'
+import { Observable } from 'rxjs'
+import { map as mapObs } from 'rxjs/operators'
 import { Cmd } from './Cmd'
 import { Sub, none } from './Sub'
 import * as platform from './Platform'
@@ -26,7 +27,7 @@ export function program<model, msg, dom>(
   subscriptions: (model: model) => Sub<msg> = () => none
 ): Program<model, msg, dom> {
   const { dispatch, cmd$, sub$, model$ } = platform.program(init, update, subscriptions)
-  const html$ = model$.map(model => view(model))
+  const html$ = model$.pipe(mapObs(model => view(model)))
   return { dispatch, cmd$, sub$, model$, html$ }
 }
 
