@@ -1,16 +1,11 @@
-import { Observable } from 'rxjs/Observable'
-import 'rxjs/add/observable/of'
-import 'rxjs/add/operator/map'
-import 'rxjs/add/operator/skip'
-import 'rxjs/add/operator/take'
-import { Subject } from 'rxjs/Subject'
-import { Task } from 'fp-ts/lib/Task'
+import { Subject, of } from 'rxjs'
 import { none as optionNone } from 'fp-ts/lib/Option'
 import { Cmd } from './Cmd'
 import { Sub, none, batch } from './Sub'
 import * as html from './Html'
 import { Location as HistoryLocation } from 'history'
 import createHashHistory from 'history/createHashHistory'
+import { map } from 'rxjs/operators'
 
 const history = createHashHistory()
 
@@ -27,12 +22,10 @@ history.listen(location => {
 })
 
 export function push<msg>(url: string): Cmd<msg> {
-  return Observable.of(
-    new Task(() => {
-      history.push(url)
-      return Promise.resolve(optionNone)
-    })
-  )
+  return of(() => {
+    history.push(url)
+    return Promise.resolve(optionNone)
+  })
 }
 
 export function program<model, msg, dom>(
@@ -42,7 +35,7 @@ export function program<model, msg, dom>(
   view: (model: model) => html.Html<dom, msg>,
   subscriptions: (model: model) => Sub<msg> = () => none
 ): html.Program<model, msg, dom> {
-  const onChangeLocation$ = location$.map(location => locationToMessage(location))
+  const onChangeLocation$ = location$.pipe(map(location => locationToMessage(location)))
   const subs = (model: model): Sub<msg> => batch([subscriptions(model), onChangeLocation$])
   return html.program(init(getLocation()), update, view, subs)
 }
