@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Observable'
+import { Observable } from 'rxjs'
 import { ReactElement } from 'react'
 import { Cmd } from './Cmd'
 import { Sub } from './Sub'
